re-org code restore some lost a11y fixes add Lap by lap data table
