+++ src/constructors/HistoryChart.tsx
@@ -93,7 +93,7 @@ export default function HistoryChart({seasons}: HistoryChartProps) {
 	
 	
 	return (
-		<Box sx={{height: 132, width: '100%'}} aria-hidden={true}>
+		<Box sx={{height: 132, width: '100%'}} aria-hidden>
 			<ResponsiveLine
 				data={[points, results, wins]}
 				colors={({color}) => color || 'transparent'}
