+++ src/drivers/CareerChart.tsx
@@ -94,7 +94,7 @@ export default function CareerChart({seasons}: CareerChartProps) {
 	
 	
 	return (
-		<Box sx={{height: 132, width: '100%'}} aria-hidden={true}>
+		<Box sx={{height: 132, width: '100%'}} aria-hidden>
 			<ResponsiveLine
 				data={[points, results, wins]}
 				colors={({color}) => color || 'transparent'}
