+++ src/race/lapTimes/LapTooltip.tsx
@@ -2,10 +2,10 @@ import {faSquare} from '@fortawesome/free-solid-svg-icons';
 import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
 import {alpha, Box, Grid, List, ListItem, ListItemAvatar, ListItemText} from '@mui/material';
 import {TooltipProps} from '@nivo/heatmap';
-import ByLine from '../drivers/ByLine';
-import DriverAvatar from '../drivers/DriverAvatar';
-import {useInvertedTheme} from '../ui-components/Theme';
-import {LapChartDatum} from './LapTimes';
+import ByLine from '../../drivers/ByLine';
+import DriverAvatar from '../../drivers/DriverAvatar';
+import {useInvertedTheme} from '../../ui-components/Theme';
+import {LapChartDatum} from './useLapTimeChartData';
 
 export default function LapTooltip({cell}: TooltipProps<LapChartDatum>) {
 	const {color, data}    = cell;
@@ -28,7 +28,7 @@ export default function LapTooltip({cell}: TooltipProps<LapChartDatum>) {
 		<List dense>
 			<ListItem>
 				<ListItemAvatar><DriverAvatar id={driverId} size={42}/></ListItemAvatar>
-				<ListItemText primary={<ByLine id={driverId} variant="name"/>} secondary={
+				<ListItemText secondaryTypographyProps={{component: 'div'}} primary={<ByLine id={driverId} variant="name"/>} secondary={
 					<Grid container spacing={2} alignItems="center">
 						<Grid item><FontAwesomeIcon icon={faSquare} color={color}/></Grid>
 						<Grid item xs>Lap {data.x}</Grid>
